+++ lib/src/enums/branta_server_base_url.dart
@@ -1,0 +1,8 @@
+enum BrantaServerBaseUrl {
+  staging('https://staging.guardrail.branta.pro'),
+  production('https://guardrail.branta.pro'),
+  localhost('http://localhost:3000');
+
+  const BrantaServerBaseUrl(this.url);
+  final String url;
+}