+++ lib/src/v2/interfaces/i_aes_encryption.dart
@@ -1,0 +1,4 @@
+abstract class IAesEncryption {
+  Future<String> encrypt(String value, String secret, {bool deterministicNonce = false});
+  Future<String> decrypt(String encryptedValue, String secret);
+}