+++ lib/src/v2/interfaces/i_secret_generator.dart
@@ -1,0 +1,4 @@
+abstract class ISecretGenerator {
+  String generate();
+  bool get deterministicNonce;
+}