+++ lib/src/v2/models/payments_result.dart
@@ -1,0 +1,8 @@
+import 'payment.dart';
+
+class PaymentsResult {
+  final List<Payment> payments;
+  final String verifyUrl;
+
+  const PaymentsResult({required this.payments, required this.verifyUrl});
+}